:truck: (interface): Rename file interface/CoreLCDDriver.hpp to interface/LCDDriver.hpp
